Refactoring form and success states into their own folders
