--- site/pages/home/contact-us-slice/style.css
@@ -1,74 +1,0 @@
-@value white, black, cainYellow, cainYellowFaded, daveGreen from "../../../css/_colors.css";
-@value mediumScreen, maxWidth from "../../../css/_sizes.css";
-
-.formContainer, .successContainer, .fatalErrorContainer {
-  padding: 30px 20px 40px 20px;
-  text-align: center;
-}
-
-.formContainer, .fatalErrorContainer {
-  composes: dividerYellow from "../../../css/_divider.css";
-  background-color: cainYellow;
-}
-
-.successContainer {
-  composes: dividerGreen from "../../../css/_divider.css";
-  background-color: daveGreen;
-}
-
-.contactUsForm, .heading {
-  max-width: maxWidth;
-}
-
-.contactUsForm {
-  margin: 0px auto;
-}
-
-.heading {
-  composes: fontL from "../../../css/typography/_fonts.css";
-  composes: serif from "../../../css/typography/_fonts.css";
-  margin: 0px auto 40px auto;
-  color: black;
-}
-
-.formLabel {
-  display: block;
-  text-align: left;
-  composes: fontS from "../../../css/typography/_fonts.css";
-  composes: sansSerif from "../../../css/typography/_fonts.css";
-  margin: 0px 12px 5px 12px;
-}
-
-.errorMessage {
-  composes: boldSansSerif from "../../../css/typography/_fonts.css";
-  display: inline-block;
-  margin-left: 10px;
-}
-
-.inputBox {
-  display: block;
-  width: 100%;
-  max-width: 100%;
-  margin-bottom: 20px;
-  composes: fontS from "../../../css/typography/_fonts.css";
-  composes: sansSerif from "../../../css/typography/_fonts.css";
-
-  padding: 12px;
-  background: cainYellowFaded;
-  border: 0px;
-  border-bottom: dotted 2px black;
-}
-
-.hasErrors {
-  border: 2px solid black;
-}
-
-.button {
-  composes: primaryButton from "../../../css/_links.css";
-  background-color: black;
-}
-
-.closeButton {
-  display: block;
-  margin: 40px auto 30px auto;
-}
